+++ lib/executor/src/request_context/web.rs
@@ -1,10 +1,10 @@
 use ntex::web;
 
-use crate::request_context::{RequestContextError, SharedRequestContext};
+use super::domains::SharedRequestContext;
+use super::error::RequestContextError;
 
 pub trait RequestContextExt {
     fn read_request_context(&self) -> Result<SharedRequestContext, RequestContextError>;
-    fn has_request_context(&self) -> bool;
     fn write_request_context(&mut self, context: SharedRequestContext);
 }
 
@@ -15,11 +15,6 @@ impl RequestContextExt for web::HttpRequest {
             .get::<SharedRequestContext>()
             .cloned()
             .ok_or(RequestContextError::Missing)
-    }
-
-    #[inline]
-    fn has_request_context(&self) -> bool {
-        self.extensions().get::<SharedRequestContext>().is_some()
     }
 
     #[inline]
@@ -35,11 +30,6 @@ impl RequestContextExt for web::WebRequest<web::DefaultError> {
             .get::<SharedRequestContext>()
             .cloned()
             .ok_or(RequestContextError::Missing)
-    }
-
-    #[inline]
-    fn has_request_context(&self) -> bool {
-        self.extensions().get::<SharedRequestContext>().is_some()
     }
 
     #[inline]
